# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaText Pre-Alpha
  description: |
    Connect with your friends effortlessly using WASAText! Send and receive messages, whether one-on-one
    or in groups, all from the convenience of your PC. Enjoy seamless conversations with text, emojis, or GIFs.
  version: "0.1"
...
servers:
  - url: https://127.0.0.1:3000
components:
  parameters:
    sourceChatId:
      name: sourceChatId
      in: path
      required: true
      schema:
        type: string
        example: "sourceChat123"
        pattern: "^[a-zA-Z0-9_-]+$"
        minLength: 1
        maxLength: 64
      description: "Unique identifier of the source chat"
...
    isGroupChat:
      name: isGroupChat
      in: query
      required: true
      schema:
        type: boolean
      description: "Indicates if the chat is a group chat"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
paths:
...
  /chats:
    get:
      tags: ["chats"]
      summary: "Retrieve user conversations"
      operationId: getMyConversations
      responses:
        "200":
          description: "Conversations retrieved successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      type: object
                      properties:
                        identifier:
                          type: string
                          description: "Unique conversation ID"
                        name:
                          type: string
                          description: "Chat or group name"
                        participants:
                          type: array
                          items:
                            type: object
                            properties:
                              userId:
                                type: string
                              username:
                                type: string
                        lastMessage:
                          type: string
                        lastMessageTimestamp:
                          type: string
                          format: date-time
                        isGroupChat:
                          type: boolean
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unable to update the photo due to a server error."
  /chats/{sourceChatId}:
    get:
      tags: ["chats"]
      summary: "Retrieve specific chat details"
      operationId: getConversation
      parameters:
        - $ref: '#/components/parameters/sourceChatId'
      responses:
        "200":
          description: "Chat details retrieved successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier:
                    type: string
                  name:
                    type: string
                  participants:
                    type: array
                    items:
                      type: object
                      properties:
                        userId:
                          type: string
                        username:
                          type: string
                  lastMessage:
                    type: string
                  lastMessageTimestamp:
                    type: string
                    format: date-time
                  isGroupChat:
                    type: boolean
        "404":
          description: "Chat not found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    pattern: "^[a-zA-Z0-9_-]+$"
                    example: "Chat not found"      
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unable to update the photo due to a server error."
...
  /chats/{sourceChatId}/Fmessages/{targetChatId}:
    post:
      tags: ["chats"]
      summary: "Forward a message from a specific chat to another specific chat"
      description: |
        Allows a user to forward a message from a specified source chat to a specified target chat.
      operationId: forwardMessage
      parameters:
        - name: sourceChatId
          in: path
          required: true
          schema:
            type: string
            example: "sourceChat123"
          description: "Unique identifier of the source chat"
        - name: targetChatId
          in: path
          required: true
          schema:
            type: string
            example: "targetChat456"
          description: "Unique identifier of the target chat"
      requestBody:
        description: "Details of the message to be forwarded"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messageId:
                  type: string
                  description: "Unique identifier of the message being forwarded"
                  example: "msg789"
                senderId:
                  type: string
                  description: "Unique identifier of the sender forwarding the message"
                  example: "user123"
                originalTimestamp:
                  type: string
                  format: date-time
                  description: "Timestamp of when the original message was sent, in ISO 8601 format"
                  example: "2023-11-01T12:00:00Z"
              required:
                - messageId
                - senderId
      responses:
        "201":
          description: "Message successfully forwarded"
          content:
            application/json:
              schema:
                type: object
                properties:
                  forwardedMessageId:
                    type: string
                    description: "Unique identifier of the forwarded message in the target chat"
                    example: "forwardedMsg101"
                  status:
                    type: string
                    description: "Status of the message forwarding action"
                    example: "Forwarded"
...
  /chats/{sourceChatId}/Grouphoto:
    post:
      tags: ["chats"]
      summary: "Change the group photo"
      operationId: setGroupPhoto
      parameters:
        - $ref: '#/components/parameters/sourceChatId'
        - $ref: '#/components/parameters/isGroupChat'
      requestBody:
        required: true
        content:
          multipart/form-data: 
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
                  description: "The new group photo to be uploaded"
                description:
                  type: string
                  example: "This is our updated group photo."
      responses:
        "200":
          description: "Group photo successfully updated"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "Group photo updated successfully"
                  photoUrl:
                    type: string
                    format: uri
                    example: "https://example.com/images/group-photo.jpg"
        "400":
          description: "Invalid file format"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid file format. Only JPEG and PNG are allowed."
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unable to update the photo due to a server error."
  /chats/{sourceChatId}/Groupname:
    post:
      tags: ["chats"]
      summary: "Change the group name"
      operationId: setGroupName
      parameters:
        - $ref: '#/components/parameters/sourceChatId'
        - $ref: '#/components/parameters/isGroupChat'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: "New group name"
      responses:
        "200":
          description: "Group name successfully updated"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "Group name updated successfully"
        "400":
          description: "Invalid name format"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid name format"
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unable to update the photo due to a server error."

### SETTINGS SECTION ###
...